Move journal abbreviation tests from BibEntryWriterTest to JournalAbbreviationRepositoryTest
